src/requirements.txt
matplotlib~=3.9.2
pymongo~=4.8.0
dash~=2.18.1
plotly~=5.24.1
pyserial~=3.5